CNDB-1707: Port remote file system abstractions to Stargazer (#273)
* LogTransaction: add ILogTransactionsFactory to provide custom log
transaction

* UCS: Port CNDB-2134 to disable shards on UCS L0

* UCS: add CompactionAggregatePrioritizer to prioritize sstables based on remote file cache

* NativeLibrary: Add INativeLibrary interface to provide custom implementation

* SSTableWatcher: to discover custom component before opening sstables

* StorageProvider: support custom file system and change Descriptor to use URI

* StorageFeatureFlags: disable features that are not supported by custom file system

* StorageHandler: to reload sstable from custom file system

(cherry picked from commit e27ee69bb404ba4c21b65d2aef157ac54425f2ef)
(cherry picked from commit e98d05a294ad2d62d1b46dad081b5f60a16b7064)
(cherry picked from commit 493bf4cb1236f6cdf819e8acbdbcf1f04d0573c5)
(cherry picked from commit 2932216864ade578d487420efa88ee17ac01a0ec)
(cherry picked from commit da98fddb2eccf91b389dd4a21976e9f81ed7b764)
(cherry picked from commit ad3ee02e41cab1089f6e4ad088fff814ca95556c)

CNDB-1707 Fix rebase compile issues

CNDB-1707 Add baseFileURI field to Descriptor for CNDB

CNDB-1707 Fix VirtualTableFromInternodeTest; cannot open virtual keyspaces in ReadCommandVerbHandler.doVerb when checking if reads can be serviced.

+++ src/java/org/apache/cassandra/io/util/FileUtils.java
@@ -51,8 +51,8 @@ import java.util.Set;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.util.concurrent.RateLimiter;
 import org.slf4j.Logger;
@@ -235,12 +235,13 @@ public final class FileUtils
     {
         try
         {
-            Files.copy(from.toPath(), to.toPath());
+            if (from.exists())
+                Files.copy(from.toPath(), to.toPath());
         }
         catch (IOException e)
         {
             if (logger.isTraceEnabled())
-                logger.trace("Could not copy file" + from + " to " + to, e);
+                logger.trace("Could not copy file " + from + " to " + to, e);
         }
     }
 
@@ -267,7 +268,11 @@ public final class FileUtils
 
     public static void truncate(String path, long size)
     {
-        File file = new File(path);
+        truncate(new File(path), size);
+    }
+
+    public static void truncate(File file, long size)
+    {
         try (FileChannel channel = file.newReadWriteChannel())
         {
             channel.truncate(size);
@@ -799,15 +804,15 @@ public final class FileUtils
      * @param source the directory containing the files to move
      * @param target the directory where the files must be moved
      */
-    public static void moveRecursively(Path source, Path target) throws IOException
+    public static void moveRecursively(File source, File target) throws IOException
     {
         logger.info("Moving {} to {}" , source, target);
 
-        if (Files.isDirectory(source))
+        if (source.isDirectory())
         {
-            Files.createDirectories(target);
+            target.tryCreateDirectories();
 
-            for (File f : new File(source).tryList())
+            for (File f : source.tryList())
             {
                 String fileName = f.name();
                 moveRecursively(source.resolve(fileName), target.resolve(fileName));
@@ -817,41 +822,60 @@ public final class FileUtils
         }
         else
         {
-            if (Files.exists(target))
+            if (target.exists())
             {
                 logger.warn("Cannot move the file {} to {} as the target file already exists." , source, target);
             }
             else
             {
-                Files.copy(source, target, StandardCopyOption.COPY_ATTRIBUTES);
-                Files.delete(source);
+                source.copy(target, StandardCopyOption.COPY_ATTRIBUTES);
+                source.delete();
             }
         }
+    }
+
+    @VisibleForTesting
+    /** @deprecated See CNDB-1707 */
+    @Deprecated(since = "5.0")
+    public static void moveRecursively(Path source, Path target) throws IOException
+    {
+        moveRecursively(new File(source), new File(target));
     }
 
     /**
      * Deletes the specified directory if it is empty
      *
-     * @param path the path to the directory
+     * @param file the path to the directory
      */
-    public static void deleteDirectoryIfEmpty(Path path) throws IOException
+    public static void deleteDirectoryIfEmpty(File file) throws IOException
     {
-        Preconditions.checkArgument(Files.isDirectory(path), String.format("%s is not a directory", path));
+        Preconditions.checkArgument(file.isDirectory(), String.format("%s is not a directory", file));
 
         try
         {
-            logger.info("Deleting directory {}", path);
-            Files.delete(path);
+            logger.info("Deleting directory {}", file);
+            Files.delete(file.toPath());
         }
         catch (DirectoryNotEmptyException e)
         {
-            try (Stream<Path> paths = Files.list(path))
-            {
-                String content = paths.map(p -> p.getFileName().toString()).collect(Collectors.joining(", "));
-
-                logger.warn("Cannot delete the directory {} as it is not empty. (Content: {})", path, content);
-            }
+            String content = Arrays.stream(file.tryList()).map(File::name).collect(Collectors.joining(", "));
+            logger.warn("Cannot delete the directory {} as it is not empty. (Content: {})", file, content);
         }
+    }
+
+    @VisibleForTesting
+    /** @deprecated See CNDB-1707 */
+    @Deprecated(since = "5.0")
+    public static void deleteDirectoryIfEmpty(Path path) throws IOException
+    {
+        deleteDirectoryIfEmpty(new File(path));
+    }
+
+    /** @deprecated See CNDB-1707 */
+    @Deprecated(since = "5.0")
+    public static long size(Path path)
+    {
+        return PathUtils.size(path);
     }
 
     public static int getBlockSize(File directory)
